accounts/pyproject.toml
[tool.aerich]
tortoise_orm = "accounts.tortoise_conf.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.poetry]
name = "accounts"
version = "0.1.0"
description = ""
authors = ["Артём Поташов <artem.potashov@bk.ru>"]

[tool.poetry.dependencies]
jwt = "^1.3.1"
bcrypt = "^4.0.1"
fastapi-mail = "^1.4.1"
pre-commit = "^2.20.0"
python = "^3.11"
fastapi = "^0.108.0"
tortoise-orm = "^0.20.0"
uvicorn = "^0.25.0"
asyncpg = "^0.29.0"
aiokafka = "^0.10.0"
aerich = "^0.7.2"
pyjwt = "^2.8.0"
fastapi-users = "^12.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
